requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --annotation-style=line --strip-extras requirements.in
#
annotated-types==0.7.0    # via pydantic
pydantic==2.10.4          # via -r requirements.in
pydantic-core==2.27.2     # via pydantic
pytz==2024.2              # via -r requirements.in
pyyaml==6.0.2             # via -r requirements.in
typing-extensions==4.12.2  # via pydantic, pydantic-core
